Inventories/requirements.txt
setuptools>=60.5.0
build>=0.7.0
packaging==21.3
wheel==0.37.1
ansible==5.4.0
importlib==1.0.4
pep517==0.12.0
tomli==2.0.1
pyparsing==3.0.7
cryptography==36.0.1
resolvelib==0.5.4
pip==22.0.3
Jinja2==3.0.3
PyYAML==6.0